fix(migrate): atomic schema migration via redis pipeline
- factor migration body out of migrate() into _migrate_into_pipeline()
  that writes through an arbitrary writer (redis client or pipeline)
- _maybe_migrate uses redis.pipeline(transaction=True) to land per-
  record rewrites and the schema_version SET in one MULTI/EXEC
- falls back to non-transactional path if pipeline() is unavailable
- move RedisDBMigrationTests to tests/test_migration.py as
  TestSchemaMigration (better co-location with cross-namespace
  migration tests)
